requirements.txt
anthropic>=0.40.0
sentence-transformers>=3.0.0
neo4j>=5.20.0
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.40.0
streamlit>=1.35.0
streamlit-agraph>=0.0.45
